refactor: 날짜 정렬 단순화화
